requirements.txt
easyocr==1.7.2
nltk==3.9.4
numpy==2.4.4
opencv_python_headless==4.13.0.92
pandas==3.0.2
Pillow==12.2.0
pytesseract==0.3.13
streamlit==1.56.0
tldextract==5.3.1
scikit-learn==1.3.0
xgboost==1.7.5
joblib==1.3.2